requirements.txt
httpx>=0.27
pydantic>=2.6
pyyaml>=6.0
typer>=0.12
rich>=13.7
pytest>=8.0
pytest-asyncio>=0.23
pdfplumber>=0.11
python-dotenv>=1.0
sarvamai>=0.1
tenacity>=8.2
fastapi>=0.110.0
uvicorn>=0.28.0
